backend/requirements.txt
asgiref==3.3.0
coverage==5.3
django==3.1.3
djangorestframework==3.12.1
pytz==2020.4
sqlparse==0.4.1
